requirements.txt
torch
numpy
matplotlib
tensorboard
notebook
deap
pynput
opencv-python
PyYAML
nbconvert
PyAV